SettingsLib: wifi: Fixed AccessPoint tracking for Passpoint networks
Previously, we were using provider name as the "SSID" for
Passpoint networks, and use it to compare against real SSIDs,
which obvious doesn't match.  So fix it.

Also removed the Passpoint specific logics in WifiTracker for
tracking Passpoint networks, since there shouldn't any
difference in tracking Passpoint APs and normal APs.

Additional work is still needed for managing installed
Passpoint configurations in the "Saved networks" page,
and it is being tracked by b/34207710. The Passpoint configuration
is not an "Access Point" per se, it is a network profile.

Bug: 34455883
Test: Install a Globla Reach Passpoint profile on a bullhead,
      verify device auto connects to a Global Reach AP and UI
      (QS, QS+, and Settings) correctly reflects it.

Change-Id: I63e66f683f8cb96e664516b130f5e7bc02358a10
Merged-In: I63e66f683f8cb96e664516b130f5e7bc02358a10

Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>