Fix client holding in fingerprint removal
When removing a fingerprint, after HAL successfully delete it,
HAL notify fingerId, and normally this fingerId is not 0.

Before this patch, RemovalClient will hold client until HAL
reporting fingerId=0 is deleted. But for vendor HAL, it is
not necessary to report fingerId=0.

With this patch, vendor HAL no need report fingerId=0 in removal,
and FingerprintService can release client in handleRemoved() so
next fingerprint request can be new client without pending.

Signed-off-by: hecong <hecong@xiaomi.com>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>